refactor: combine LitterBox analyzer services
