+++ src/main/java/de/uni_passau/fim/se2/litterbox_web/linter/LinterController.java
@@ -9,57 +9,36 @@
  */
 package de.uni_passau.fim.se2.litterbox_web.linter;
 
-import java.util.List;
+import java.util.Locale;
+import java.util.Set;
 
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import de.uni_passau.fim.se2.litterbox.ast.model.Program;
-import de.uni_passau.fim.se2.litterbox_web.shared.Scratch3ParserService;
+import de.uni_passau.fim.se2.litterbox_web.shared.LitterBoxAnalysisService;
 import de.uni_passau.fim.se2.litterbox_web.shared.dto.IssueDTO;
 
 @RestController
 @RequestMapping("/linter")
 public class LinterController {
 
-    private final Scratch3ParserService parserService;
+    private final LitterBoxAnalysisService litterBoxAnalysisService;
 
-    private final LinterService linterService;
-
-    public LinterController(
-        final Scratch3ParserService parserService,
-        final LinterService linterService
-    ) {
-        this.parserService = parserService;
-        this.linterService = linterService;
+    public LinterController(final LitterBoxAnalysisService litterBoxAnalysisService) {
+        this.litterBoxAnalysisService = litterBoxAnalysisService;
     }
 
     /**
-     * Analyses a Scratch 3 program using LitterBox and returns the found issues.
+     * Analyses the program from the request and returns the generated feedback
      *
-     * @param projectJson A Scratch 3 program in json file format.
-     * @param locale      Language/locale used for analysis
-     * @param detectors   Programm analyzer detectors for filtering found issues.
-     * @return The found LitterBox issues.
+     * @param request the API request converted to {@link LinterRequest}.
+     * @return feedback for the submitted program as string.
      */
     @PostMapping("analyze")
-    public List<IssueDTO> analyze(
-        @RequestParam(
-            value = "locale",
-            required = false,
-            defaultValue = "en"
-        ) String locale,
-        @RequestParam(
-            value = "detectors",
-            required = false,
-            defaultValue = "all"
-        ) String detectors,
-        @RequestBody String projectJson
-    ) {
-        final Program program = parserService.parseFromString(projectJson);
-        return linterService.getIssues(program, locale, detectors);
+    public Set<IssueDTO> generateFeedback(@RequestBody final LinterRequest request) {
+        final Locale language = request.languageAsLocale();
+        return litterBoxAnalysisService.getIssues(request.program(), language, request.detectors());
     }
 }
