--- src/main/java/de/uni_passau/fim/se2/litterbox_web/tutorial_system/TutorialIssueInfo.java
@@ -1,22 +1,0 @@
-/*
- * Copyright (C) 2024 LitterBox-Web contributors
- *
- * This file is part of LitterBox-Web.
- * Licenced under the EUPL-1.2 or later.
- *
- * SPDX-FileCopyrightText: 2024 LitterBox-Web contributors
- * SPDX-License-Identifier: EUPL-1.2
- */
-package de.uni_passau.fim.se2.litterbox_web.tutorial_system;
-
-import de.uni_passau.fim.se2.litterbox.analytics.IssueType;
-
-public record TutorialIssueInfo(
-    String name,
-    String hint,
-    String sprite,
-    String costume,
-    IssueType type,
-    String scratchBlocksCode
-) {
-}
